+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/logs/data-streams/data-streams.component.js
@@ -1,0 +1,15 @@
+import controller from './data-streams.controller';
+import template from './data-streams.template.html';
+
+export default {
+  bindings: {
+    projectId: '<',
+    region: '<',
+    loadbalancerId: '<',
+    kind: '<',
+    goBack: '<',
+    trackClick: '<',
+  },
+  controller,
+  template,
+};