+++ packages/manager/modules/log-to-customer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-log-to-customer@2.0.0...@ovh-ux/manager-log-to-customer@2.1.0) (2024-07-04)
+
+
+### Bug Fixes
+
+* **i18n:** add missing translations [CDS 3131] ([#12227](https://github.com/ovh/manager/issues/12227)) ([039bb75](https://github.com/ovh/manager/commit/039bb75a8a524e046cd1e8a1d69c7b18f7177730))
+
+
+### Features
+
+* add Logs to customer for Octavia Load Balancer ([#11860](https://github.com/ovh/manager/issues/11860)) ([0d66ac3](https://github.com/ovh/manager/commit/0d66ac3cbeccf4aa3c9464c08230077f1649c231))
+
+
+
+
+
 ## [1.1.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-log-to-customer@1.1.0...@ovh-ux/manager-log-to-customer@1.1.1) (2024-06-25)